Add processing of incoming text messages
Searches for a vocab entry for the incoming text in the local vocab repo.
If it is not found there then calls Yandex.Dictionary API and
saves the result to the local repo.
Then replies to user with a short description of the vocab entry and the button
with a callback command for showing a full description.
+ refactoring